pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "echelonos"
version = "0.1.0"
description = "Automated contract obligation extraction pipeline with LLM-powered analysis and evidence trails"
requires-python = ">=3.11"
dependencies = [
    # Orchestration
    "prefect>=3.0",
    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    # LLM Clients
    "openai>=1.50",
    "anthropic>=0.40",
    # OCR
    "azure-ai-documentintelligence>=1.0",
    # Document processing
    "pypdf>=4.0",
    "python-docx>=1.1",
    "python-magic>=0.4",
    # Hashing / dedup
    "simhash>=2.1",
    # Config
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # Utilities
    "structlog>=24.0",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.hatch.build.targets.wheel]
packages = ["src/echelonos"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "e2e: end-to-end tests",
    "unit: unit tests",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
